Rename insertActions to InsertLocation
